Moved the traits and implementations into separate files and removed the ImplementedBY annotations
