+++ server/app/uk/gov/ons/addressIndex/server/modules/AddressParserModule.scala
@@ -1,22 +1,8 @@
 package uk.gov.ons.addressIndex.server.modules
 
 import javax.inject.{Inject, Singleton}
-
-import com.google.inject.ImplementedBy
 import uk.gov.ons.addressIndex.crfscala.CrfScalaJniImpl
 import uk.gov.ons.addressIndex.parsers.Parser
-
-@ImplementedBy(classOf[AddressParserModule])
-trait ParserModule {
-  /**
-    * Transforms input string into labeled address modules
-    * Uses JNI and a generated (by datascientists) model of CRF (conditional random fields)
-    *
-    * @param input string to be tokenized
-    * @return List of labeled tokens
-    */
-  def parse(input: String): Map[String, String]
-}
 
 @Singleton
 class AddressParserModule @Inject()(conf: ConfigModule) extends ParserModule {
@@ -33,7 +19,6 @@ class AddressParserModule @Inject()(conf: ConfigModule) extends ParserModule {
 
     new Parser(native)
   }
-
 
   def parse(input: String): Map[String, String] = parser.parse(input)
 }
