+++ server/app/uk/gov/ons/addressIndex/server/modules/SystemBootstrap.scala
@@ -1,9 +1,9 @@
 package uk.gov.ons.addressIndex.server.modules
 
 import java.io.File
-import javax.inject.{Inject, Singleton}
 
 import com.typesafe.config.ConfigFactory
+import javax.inject.{Inject, Singleton}
 import play.api.Logger
 
 @Singleton
