+++ server/app/uk/gov/ons/addressIndex/server/utils/HopperScoreHelper.scala
@@ -3,7 +3,6 @@ package uk.gov.ons.addressIndex.server.utils
 import uk.gov.ons.addressIndex.model.db.BulkAddress
 import uk.gov.ons.addressIndex.model.server.response.AddressResponseAddress
 import uk.gov.ons.addressIndex.parsers.Tokens
-import uk.gov.ons.addressIndex.server.utils.impl.GenericLogger
 
 import scala.util.Try
