+++ server/test/uk/gov/ons/addressIndex/server/controllers/AddressControllerSpec.scala
@@ -13,8 +13,7 @@ import uk.gov.ons.addressIndex.model.db.{BulkAddress, BulkAddressRequestData, Bu
 import uk.gov.ons.addressIndex.model.server.response._
 import uk.gov.ons.addressIndex.server.modules._
 import uk.gov.ons.addressIndex.server.modules.validation._
-import uk.gov.ons.addressIndex.server.utils.HopperScoreHelper
-import uk.gov.ons.addressIndex.server.utils.impl.APIThrottle
+import uk.gov.ons.addressIndex.server.utils.{APIThrottle, HopperScoreHelper}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration.Duration
