--- version.sbt
@@ -1,1 +1,0 @@
-version in ThisBuild := "0.0.1"
